FIN-28 Receivables refactoring - StickyTable reusable component
