+++ src/src_V2/components/confirm-dialog/ConfirmDialog.jsx
@@ -32,6 +32,7 @@ export default function ConfirmDialog({
   const cancelBtn = {
     variant: 'outlined',
     txt: 'Cancel',
+    focus: false,
     ...cancelButton,
   };
 
@@ -58,7 +59,7 @@ export default function ConfirmDialog({
       <DialogActions>
         <Button
           onClick={cancelHandler}
-          autoFocus={!confirmBtn.focus}
+          autoFocus={cancelBtn.focus || !confirmBtn.focus}
           variant={cancelBtn.variant}
         >
           {cancelBtn.txt}
