+++ src/src_V2/modules/manager/financial/components/index.js
@@ -1,2 +1,7 @@
 export { default as SearchInput } from './search-input/SearchInput';
 export { default as DateInput } from './date-input/DateInput';
+export {
+  default as StickyTable,
+  TooltipTableCell,
+  TableCellWrapper,
+} from './sticky-table/StickyTable';
