+++ src/src_V2/modules/manager/financial/constants/index.js
@@ -5,6 +5,7 @@ export {
   paymentHistoryColumns,
   paymentHistoryColumnTypes,
   payableValidationValues,
+  payableColumnTypes,
 } from './payable-table';
 export {
   receivableColumns,
