+++ src/src_V2/modules/manager/financial/payables/table/custom-cells/CustomCells.module.css
@@ -1,0 +1,25 @@
+/* Table Cell Statuses */
+.statusChip {
+  height: 27px !important;
+}
+
+.canceledPayableStatus {
+  background-color: var(--midLightGrey) !important;
+  color: var(--secondary) !important;
+  border: 1px solid var(--secondary) !important;
+}
+
+/* todo duplication */
+.whiteTableCell {
+  background-color: var(--white) !important;
+}
+
+.tableCell {
+  min-width: 140px !important;
+}
+
+/* Attachment */
+.attachmentImg {
+  margin-top: 10px;
+  color: var(--secondary);
+}
