+++ src/src_V2/modules/manager/financial/utils/table.js
@@ -1,0 +1,14 @@
+// todo find another solution for style import
+import styles from '../components/sticky-table/StickyTable.module.css';
+
+export function colorCondition(difference) {
+  let style;
+  if (difference > 0) {
+    style = styles.positiveTableCell;
+  } else if (difference < 0) {
+    style = styles.negativeTableCell;
+  } else {
+    style = styles.neutralTableCell;
+  }
+  return style;
+}
